removed warnings and minor issues

+++ include/anira/InferenceConfig.h
@@ -71,13 +71,13 @@ struct ANIRA_API InferenceConfig {
         if (m_model_input_shape_onnx.size() > 0) {
             m_new_model_input_size = 1;
             for (int i = 0; i < m_model_input_shape_onnx.size(); ++i) {
-                m_new_model_input_size *= m_model_input_shape_onnx[i];
+                m_new_model_input_size *= (int) m_model_input_shape_onnx[i];
             }
         }
         if (m_model_output_shape_onnx.size() > 0) {
             m_new_model_output_size = 1;
             for (int i = 0; i < m_model_output_shape_onnx.size(); ++i) {
-                m_new_model_output_size *= m_model_output_shape_onnx[i];
+                m_new_model_output_size *= (int) m_model_output_shape_onnx[i];
             }
         }
 #elif USE_TFLITE
